feat: slack channel management (#124)
## Enhancements to Spacecat Shared Slack Client

### Overview

This PR introduces significant enhancements to the
`@adobe/spacecat-shared-slack-client` package. The focus is on added
functionalities for channel creation and user invitations in the context
of SpaceCat handling collaboration channel creation for fulfillment
events as well as when requested via HTTP API for select clients, such
as Cloudmanager.

### Main Features

1. **Channel Management**: Functionality to create new Slack channels,
with options for setting the channel's topic, description, and privacy
settings. For each created channel, configurable ops admins are invited.

2. **User Invitations**: A feature to invite users to Slack channels via
email, catering to various user states like existing users, users not in
channel, and non-workspace members, as well as users needing upgrade
from single-channel to multi-channel guests.

3. Ops config: ability to specify an ops slack channel per slack target.
This is used to post status and action required messages for human
follow-up.

### Logical Flow

- **Client Instantiation**: The flow begins with the creation of a Slack
client using the `createFrom` factory method. This method ensures the
right type of client is instantiated based on the target and privilege
level.
  
- **Operations**: Depending on the client type, operations like posting
messages, creating channels, inviting users, and uploading files are
performed. These operations make API calls to Slack, handling responses
and errors accordingly.

- **Error Handling and Logging**: The implementation includes error
handling and logging.

- **Timing Information**: Most operations and all Slack API calls are
timed and debug-logged.