feat: multi-query support for rum-client (#311)
#### Add Support for Multiple Queries in RUM Client

This PR introduces support for handling multiple queries in the RUM
client, a feature necessary for opportunity detection.

The `queryMulti` method works by first fetching the RUM bundles once and
then applying all requested queries to the retrieved raw data. This is
to make sure the RUM Bundler API is only called once.