feat: SITES-26591 [Import Assistant] Expand GPT client to user chat endpoint (#436)
My PR did not have the `feat` prefix and so it wasn't deployed. I didn't
realize that was the mechanism. Is that worth a description in
`pull_request_template.md`?

This is a quick correction to some JavaDocs but essentially is to have
the GPT package deployed.

## Related Issues
https://jira.corp.adobe.com/browse/SITES-26591
PR: https://github.com/adobe/spacecat-shared/pull/422