fix: Update shared-data schema to include 'da' type in options (#839)
To include support for da imports, we need to define type property in
our options field. The value for type can either be 'xwalk' or 'doc' or
'da'. The schema needs to be updated to support this new property.

<img width="868" alt="Screenshot 2025-07-08 at 6 31 21 PM"
src="https://github.com/user-attachments/assets/7df44017-59ab-4b00-97e8-74618fbeebd9"
/>

## Related Issues
* Parent issue for supporting DA import in EC :
https://github.com/Adobe-AEM-Foundation/site-transfer-agent-import/issues/59