Add `'never'` to job intervals
Add `'never'` as a valid value for job intervals.

This enables activating imports for sites without ever having them
scheduled by `spacecat-job-dispatcher`.

That's useful for imports that are invoked manually via the `@spacecat`
bot or triggered by audits.

For ahrefs-based imports specifically, this is meant to drive down the
consumption of Ahrefs API Units.