feat: add isOriginal field to scrapeUrl (#908)
Added an additional field to the ScrapeUrl model.
This field is used to determine if a ScrapeUrl is an original or a
skipped url that uses the scrape of an original ScrapeUrl. New
ScrapeUrls check if there is a original ScrapeUrl within their
maxScrapeAge parameter. This prevents the problem of a sliding time
window while checking for existing ScrapeUrls.