fix: add readability as audit type (#926)
added readability to AUDIT_TYPES. Please let me know if all makes sense
or if anything else is needed! Let me know if package or package-lock
needs to be updated.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes

## Related Issues


Thanks for contributing!

Co-authored-by: dogadogan <doga@adobe.com>