feat: adding new prerender audit type (#959)
**Requirement**
With **tokowaka**, we are introducing a new **prerender** opportunity
that would help customer identify the content gap b/w what is visible to
the customer vs llm and enable them to reduce this gap by onboarding on
tokowaka.

Adding new audit type in spacecat = **prerender** for same.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes

## Related Issues


Thanks for contributing!