fix: adds browser export (#1012)
Adds a browser export to the shared-utils package.
This is needed as the current package does not support treeshaking, and
it is impossible to load only browser specific functions out of the box.

This approach tells the browser which entry point provides browser
supported methods (as well as avoid bundling the rest)

## Related Issues


Thanks for contributing!