feat: start breaking dependency cycles in `shared` (#1052)
Starts breaking dependency cycles.

- Copies `DELIVERY_TYPES` to shared-utils.

This is the groundwork to eventually remove it from data-access and
breaks the dependency cycle `spacecat-shared-utils <=>
spacecat-shared-data-access`

- Copies auth helpers to shared-ims-client. 

This is the groundwork to eventually remove it from data-access and
breaking the dependency cycle `spacecat-shared-utils <=>
spacecat-shared-ims-client`. For this to happen, we have to land these
changes first, and change package dependencies afterwards.

The stepped approach is necessary because packages in this repository
don't have access to the _future_ package versions of shared packages.
This can eventually be resolved by adopting a proper workspace approach.

Please ensure your pull request adheres to the following guidelines:
- [x] make sure to link the related issues in this description
- [x] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes

## Related Issues
https://jira.corp.adobe.com/browse/LLMO-872

Thanks for contributing!