fix: use correct model name (#1199)
Use correct model name to properly remove `configurationId` prior to
creating a new config version. otherwise the following condition
violation occurs:

```
{
    "Item":
    {
        "configurationId": "3d85e3f7-f4c8-4151-9aa5-325ae43c7c0f",
        "createdAt": "2025-12-01T08:03:45.538Z",
        "updatedAt": "2025-12-01T08:03:46.595Z",
        "updatedBy": "system",
        "version": 3201,
        "versionString": "0000003201",
        "pk": "$spacecat#configurationid_3d85e3f7-f4c8-4151-9aa5-325ae43c7c0f",
        "sk": "$configuration_1",
        "gsi1pk": "all_configurations",
        "gsi1sk": "$configuration_1#versionstring_0000003201",
        "__edb_e__": "Configuration",
        "__edb_v__": "1"
    },
    "TableName": "spacecat-services-data",
    "ConditionExpression": "attribute_not_exists(#pk) AND attribute_not_exists(#sk)",
    "ExpressionAttributeNames":
    {
        "#pk": "pk",
        "#sk": "sk"
    }
}
```
this because the new config is a copy of the old one, still containing
the full ID, thus creating a "primary key" (pk + sk) collision. the copy
needs to have a new ID, which works best by removing it from the payload
and have the schema auto-generate it. this was intended, but the wrong
model name was used for sanitization.